Isse113: Refactoring InformationSheetContent file: Commit 2: 70% work done
